Group element helpers in different files
According to the section at https://developer.mozilla.org/en-US/docs/Web/HTML/Element
